Q2

+++ Question2.java
@@ -7,10 +7,10 @@ public class Question2
     Scanner in = new Scanner(System.in); 
 
     System.out.println("Enter your height (in m): ");
-    double height = in.nextDouble();
+    int height = in.nextDouble();
 
     System.out.println("Enter your weight (in kg): ");
-    double weight = in.nextDouble();
+    int weight = in.nextDouble();
 
     double BMI = weight / (height * height);
     System.out.println("Your BMI is " + BMI);
